split user.py
